Inline Eval class into certificate verifier
